moving server change logic to Layout
