+++ src/components/Layout.js
@@ -1,4 +1,4 @@
-import { useContext, useMemo } from 'react'
+import { useContext, useEffect, useMemo } from 'react'
 import PropTypes from 'prop-types'
 import { AuthenticationContext } from 'gitea-react-toolkit'
 import Header from '@components/Header'
@@ -6,14 +6,16 @@ import Footer from '@components/Footer'
 import Onboarding from '@components/Onboarding'
 import { StoreContext } from '@context/StoreContext'
 import { getBuildId } from '@utils/build'
-import { APP_NAME, QA, QA_BASE_URL } from '@common/constants'
+import { APP_NAME, BASE_URL, PROD, QA, QA_BASE_URL } from '@common/constants'
 import useValidateAccountSettings from '@hooks/useValidateAccountSettings'
+import { useRouter } from 'next/router'
 
 export default function Layout({
   children,
   showChildren,
   title = APP_NAME,
 }) {
+  const router = useRouter()
   const {
     state: authentication,
     component: authenticationComponent,
@@ -29,8 +31,24 @@ export default function Layout({
     actions: {
       setCurrentLayout,
       setShowAccountSetup,
+      setServer,
     },
   } = useContext(StoreContext)
+
+  useEffect(() => {
+    const params = router?.query
+
+    if (typeof params?.server === 'string') { // if URL param given
+      const serverID_ = params.server.toUpperCase() === QA ? QA : PROD
+      const server_ = (serverID_ === QA) ? QA_BASE_URL : BASE_URL
+
+      if (server !== server_) {
+        console.log(`_app.js - On init switching server to: ${serverID_}, url server param '${params.server}', old server ${server}, reloading page`)
+        setServer(server_) // persist server selection in localstorage
+        router.push(`/?server=${serverID_}`) // reload page
+      }
+    }
+  }, [router?.query]) // TRICKY query property not loaded on first pass, so watch for change
 
   const buildId = useMemo(getBuildId, [])
   useValidateAccountSettings(authentication, showAccountSetup, languageId, owner, setShowAccountSetup)
